Obtendo usuário logado

+++ DotnetRpg/Services/CharacterService.cs
@@ -1,10 +1,12 @@
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using DotnetRpg.Data;
 using DotnetRpg.Dtos.Character;
 using DotnetRpg.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace DotnetRpg.Services
@@ -13,16 +15,23 @@ namespace DotnetRpg.Services
     {
         private readonly IMapper _mapper;
         private readonly DataContext _context;
-        
-        public CharacterService(IMapper mapper, DataContext context)
+        private readonly IHttpContextAccessor _httpContextAccessor;
+
+        public CharacterService(IMapper mapper, DataContext context,
+            IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
-        
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+
+        private int GetUserId() => int.Parse(_httpContextAccessor
+            .HttpContext
+            .User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
-            var dbCharacters = await _context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+            var dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>
             {
                 Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList()
@@ -32,7 +41,7 @@ namespace DotnetRpg.Services
 
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             var serviceResponse = new ServiceResponse<GetCharacterDto>
             {
                 Data = _mapper.Map<GetCharacterDto>(dbCharacter)
@@ -44,9 +53,11 @@ namespace DotnetRpg.Services
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             var character = _mapper.Map<Character>(newCharacter);
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+
             await _context.AddAsync(character);
             await _context.SaveChangesAsync();
-            serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+            serviceResponse.Data = _context.Characters.Where(c => c.User.Id == GetUserId()    ).Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             return serviceResponse;
         }
 
@@ -55,7 +66,7 @@ namespace DotnetRpg.Services
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
             try
             {
-                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id && c.User.Id == GetUserId());
                 character.Name = updatedCharacter.Name;
                 character.Defense = updatedCharacter.Defense;
                 character.Class = updatedCharacter.Class;
@@ -74,8 +85,6 @@ namespace DotnetRpg.Services
                 serviceResponse.Message = ex.Message;
             }
 
-
-
             return serviceResponse;
         }
 
@@ -84,11 +93,21 @@ namespace DotnetRpg.Services
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
+                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if(character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    serviceResponse.Data = _context.Characters
+                        .Where(c => c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found";
+                }
 
-                serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             }
             catch (System.Exception ex)
             {
